refactor(clock): simplify TimerSkippingRealExecutionClock
